requirements.txt
pandas
numpy
matplotlib
sqlite3
jupyter
python-dotenv
